refactor: orderTable, TableGroup을 table 패키지로 분리
