+++ src/main/java/kitchenpos/order/application/OrderService.java
@@ -3,14 +3,14 @@ package kitchenpos.order.application;
 import kitchenpos.exception.NonExistentException;
 import kitchenpos.order.domain.Order;
 import kitchenpos.order.domain.OrderStatus;
-import kitchenpos.order.domain.OrderTable;
 import kitchenpos.order.domain.repository.OrderRepository;
-import kitchenpos.order.domain.repository.OrderTableRepository;
 import kitchenpos.order.ui.dto.OrderDetailResponse;
 import kitchenpos.order.ui.dto.OrderRequest;
 import kitchenpos.order.ui.dto.OrderResponse;
 import kitchenpos.order.ui.dto.OrderStatusRequest;
 import kitchenpos.order.ui.dto.OrderStatusResponse;
+import kitchenpos.table.domain.OrderTable;
+import kitchenpos.table.domain.repository.OrderTableRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
