+++ src/main/java/kitchenpos/table/domain/OrderStatusCheckEvent.java
@@ -1,4 +1,4 @@
-package kitchenpos.order.domain;
+package kitchenpos.table.domain;
 
 import org.springframework.context.ApplicationEvent;
