+++ src/test/java/kitchenpos/acceptance/DomainAcceptanceTest.java
@@ -12,11 +12,11 @@ import kitchenpos.menu.ui.dto.MenuUpdateRequest;
 import kitchenpos.order.ui.dto.OrderLineItemRequest;
 import kitchenpos.order.ui.dto.OrderRequest;
 import kitchenpos.order.ui.dto.OrderResponse;
-import kitchenpos.order.ui.dto.OrderTableIdRequest;
-import kitchenpos.order.ui.dto.TableGroupRequest;
-import kitchenpos.order.ui.dto.TableGroupResponse;
-import kitchenpos.order.ui.dto.TableRequest;
-import kitchenpos.order.ui.dto.TableResponse;
+import kitchenpos.table.ui.dto.OrderTableIdRequest;
+import kitchenpos.table.ui.dto.TableGroupRequest;
+import kitchenpos.table.ui.dto.TableGroupResponse;
+import kitchenpos.table.ui.dto.TableRequest;
+import kitchenpos.table.ui.dto.TableResponse;
 import kitchenpos.product.ui.dto.ProductRequest;
 import kitchenpos.product.ui.dto.ProductResponse;
 import org.springframework.http.HttpStatus;
