+++ src/test/java/kitchenpos/acceptance/TableGroupAcceptanceTest.java
@@ -3,8 +3,8 @@ package kitchenpos.acceptance;
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
-import kitchenpos.order.ui.dto.OrderTableIdRequest;
-import kitchenpos.order.ui.dto.TableGroupRequest;
+import kitchenpos.table.ui.dto.OrderTableIdRequest;
+import kitchenpos.table.ui.dto.TableGroupRequest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
