+++ src/test/java/kitchenpos/application/TableGroupServiceTest.java
@@ -1,9 +1,9 @@
 package kitchenpos.application;
 
 import kitchenpos.exception.NonExistentException;
-import kitchenpos.order.application.OrderTableService;
-import kitchenpos.order.application.TableGroupService;
-import kitchenpos.order.domain.repository.TableGroupRepository;
+import kitchenpos.table.application.OrderTableService;
+import kitchenpos.table.application.TableGroupService;
+import kitchenpos.table.domain.repository.TableGroupRepository;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.mockito.InjectMocks;
