+++ src/test/java/kitchenpos/order/OrderTableTest.java
@@ -1,7 +1,7 @@
 package kitchenpos.order;
 
-import kitchenpos.order.domain.OrderTable;
-import kitchenpos.order.domain.TableGroup;
+import kitchenpos.table.domain.OrderTable;
+import kitchenpos.table.domain.TableGroup;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
