+++ src/test/java/kitchenpos/order/OrderTest.java
@@ -3,7 +3,7 @@ package kitchenpos.order;
 import kitchenpos.exception.FieldNotValidException;
 import kitchenpos.order.domain.Order;
 import kitchenpos.order.domain.OrderStatus;
-import kitchenpos.order.domain.OrderTable;
+import kitchenpos.table.domain.OrderTable;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
