+++ src/test/java/kitchenpos/utils/Fixture.java
@@ -8,14 +8,14 @@ import kitchenpos.menu.ui.dto.MenuRequest;
 import kitchenpos.menu.ui.dto.MenuUpdateRequest;
 import kitchenpos.order.domain.Order;
 import kitchenpos.order.domain.OrderLineItem;
-import kitchenpos.order.domain.OrderTable;
-import kitchenpos.order.domain.TableGroup;
+import kitchenpos.table.domain.OrderTable;
+import kitchenpos.table.domain.TableGroup;
 import kitchenpos.order.ui.dto.OrderLineItemRequest;
 import kitchenpos.order.ui.dto.OrderRequest;
 import kitchenpos.order.ui.dto.OrderStatusRequest;
-import kitchenpos.order.ui.dto.OrderTableIdRequest;
-import kitchenpos.order.ui.dto.TableGroupRequest;
-import kitchenpos.order.ui.dto.TableRequest;
+import kitchenpos.table.ui.dto.OrderTableIdRequest;
+import kitchenpos.table.ui.dto.TableGroupRequest;
+import kitchenpos.table.ui.dto.TableRequest;
 import kitchenpos.product.domain.Product;
 import kitchenpos.product.ui.dto.ProductRequest;
